requirements.txt
imageio==2.37.0
imageio-ffmpeg==0.6.0
nibabel==5.3.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
psutil==7.0.0
typing-extensions==4.14.0
scikit-image
numpy-stl
